restructured my account page
