+++ src/pages/MyAccountPage/MyAccount.jsx
@@ -1,11 +1,11 @@
 import { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, Navigate } from "react-router-dom";
-import animation from "../animations/popupbox";
-import LoadingScreen from "../components/loadingScreen";
-import Modal from "../components/modal";
-import { apiURL } from "../store/actions";
-import { todolistSlice } from "../store/todolistSlice";
+import animation from "../../animations/popupbox";
+import LoadingScreen from "../../components/loadingScreen";
+import Modal from "../../components/modal";
+import { apiURL } from "../../store/actions";
+import { todolistSlice } from "../../store/todolistSlice";
 
 function MyAccount() {
   const currentUser = useSelector((state) => state.todos.AllUserInfo);
