+++ src/styles/index.scss
@@ -24,6 +24,7 @@ body {
 @import "./pages/ForgottenPasswordPage/verify";
 @import "./pages/UpdatePages/update";
 @import "./pages/SettingsPage/settings";
+@import "./pages/MyAccountPage/my-account";
 
 //dark/light mode switch styles
 .main-container-relative {
